Many optimizations
main render loop now 2-3x faster on my machine

+++ score.py
@@ -6,4 +6,4 @@ class Score:
 
     def __init__(self, image_path):
         self._image = Image.open(image_path)
-        self.data = numpy.asarray(self._image)
+        self.data = numpy.asarray(self._image, dtype=numpy.uint16)
